refactor(issue#64): add modifications based on folder structure discussion
